+++ lib/auxiliary_rails/application/form.rb
@@ -6,6 +6,8 @@ module AuxiliaryRails
       include ActiveModel::Attributes
       include ActiveModel::AttributeAssignment
       include ActiveModel::Model
+      include ActiveModel::Validations
+      include AuxiliaryRails::Concerns::Performable
 
       class << self
         # Method for ActiveModel::Translation
@@ -14,8 +16,13 @@ module AuxiliaryRails
         end
       end
 
-      def submit
-        raise NotImplementedError
+      # Method for ActiveModel::Errors
+      def read_attribute_for_validation(attr_name)
+        if attr_name.to_sym == :form
+          self
+        else
+          super
+        end
       end
     end
   end
